alerter/requirements.txt
torch>=2.0.0
numpy>=1.24.0
requests>=2.28.0
prometheus_client>=0.19.0
websocket-client>=1.6.0